📦 NEW: display boosted posts with an according indication
